[FLINK-38544][checkpoint] Phase 5 fix: relocate placeholder ITCases to flink-runtime same-package
The two unit-style placeholder fixtures introduced by Phase 5 (Unaligned
checkpoint-during-recovery and rescale-large-record OOM regression) reach into
SpillFile's package-private segments() / SpillFileSegment, which fails to
compile from the org.apache.flink.test.checkpointing package under flink-tests.

Move both files to flink-runtime/src/test/java/.../checkpoint/channel/ so they
sit in the same package as SpillFile and DiskSnapshot, and remove the now-unused
qualified imports. Real MiniCluster ITCases against the live recovery path
remain pending manual verification per requirements/38544/phase5_coordination/
acceptance_test.md (AT-T5QE, AT-FZXD).
